@@ -8,12 +8,13 @@
     "lint": "eslint *.ts*"
   },
   "devDependencies": {
-    "@types/react": "^17.0.37",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "18.0.25",
+    "@types/react-dom": "18.0.9",
     "eslint": "^7.32.0",
     "eslint-config-custom": "*",
-    "react": "^17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "tsconfig": "*",
-    "typescript": "^4.5.2"
+    "typescript": "4.9.3"
   }
 }
